Some fixes and improvement of the Android makefile and gitignore
RAYLIB_PATH: Hard path replaced by a relative path matching the description in the README.
ANDROID_API_VERSION: Can now be set with the make command.
clear: When running the make command a 2nd time, the APK building never stopped, with its size increasing forever. Added a clear step to remove the bin folder, so building can go normally.
create_project_apk_package: Updated temp APK name so we use the correct one for the rest of the process.
Added "*apk*" and "android.*" to the gitignore.